#33 Move array function module under root
