+++ decode/__init__.py
@@ -3,11 +3,8 @@ __version__ = "0.6.1"
 __author__ = "DESHIMA software team"
 
 
-# subpackages
-from . import core
-
-
 # submodules
+from . import array
 from . import cube
 from . import io
 from . import models
@@ -17,8 +14,8 @@ from . import utils
 
 
 # aliases
+from .array import *
 from .cube import *
-from .core import *
 from .logging import *
