+++ decode/array.py
